// generated/src/app/api-models/prov-privilege.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ProvPrivilege {
  seqProvPrivilege: number;
  seqProvId: number;
  seqHospId: number;
  effectiveDate: Date;
  expirationDate: Date;
  userDefined: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  privilegeType: string;
}